+++ Addon/ElvUI_mMediaTag/core/changelog/3.67.lua
@@ -1,7 +1,11 @@
 mMT.Changelog[367]  = {
-	DATE = "Dev not",
-	IMPORTANT = { "there is nothing new here" },
-	NEW = { "there is nothing new here" },
-	UPDATE = { "there is nothing new here" },
-	FIX = { "there is nothing new here" },
+	DATE = "07.12.2024",
+	NEW = {
+		"[Datatext-Professions] Add new Custom Icons for Professions to the professions datatexts.",
+		"[Textures] Statusbar Texture mMT Dark."
+	 },
+	FIX = {
+		"[Cosmeticbars]: Fixed Border Visibility settings if Cosmeticbars is applied to the panel.",
+		"[Portraits]: Fix bug with rare elite units and the new Border colors."
+	 },
 }